% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averaging_duplicated_genes.r
\name{avg_function}
\alias{avg_function}
\title{An expression averaging function for an unique gene}
\usage{
avg_function(gene_name, exprs)
}
\arguments{
\item{gene_name}{a character of gene name}

\item{exprs}{a matrix of normalized expression data (TPM or FPKM) containing n-rows of cells 
and m-cols of genes}
}
\value{
an expression data with average expression value computed for \code{gene_name}
}
\description{
An expression averaging function which collapses multiple expression with same gene name to the 
average value
}
\author{
MaiChan Lau
}

