% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averaging_duplicated_genes.r
\name{avg_same_genes}
\alias{avg_same_genes}
\title{An expression averaging function for duplicated genes}
\usage{
avg_same_genes(exprs = NULL)
}
\arguments{
\item{exprs}{a matrix of normalized expression data (TPM or FPKM) containing n-rows of cells 
and m-cols of genes}
}
\value{
an averaged expression data with unique gene names
}
\description{
An expression averaging function which takes the average value of duplicated genes as 
Monocle's CellDataSet object cannot handle duplicated genes.
}
\author{
MaiChan Lau
}

