% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2c_dist_uneveness_cyclical.r
\name{c2c_dist_uneveness_cyclical}
\alias{c2c_dist_uneveness_cyclical}
\title{A c2c distance uneveness function}
\usage{
c2c_dist_uneveness_cyclical(d, alpha = 0.3)
}
\arguments{
\item{d}{an expresssion matrix containing n-by-n pairwise distances computed for n cells}

\item{alpha}{window size which is given in the percentage of total no. of cells}
}
\value{
a value of total uneveness
}
\description{
A c2c distance uneveness function allows you to compute the overall cell-to-cell (c2c)
distance variability or uneveness which is summed over small local windows surrounding 
individual cells.
}
\author{
MaiChan Lau
}

