% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_monocle_functions.r
\name{clusterGenes}
\alias{clusterGenes}
\title{A modified monocle's function}
\usage{
clusterGenes(expr_matrix, krange, method = function(x) {     as.dist((1 -
  cor(t(x)))/2) }, ...)
}
\arguments{
\item{full_models}{a Monocle's vgam full model}

\item{reduced_models}{a Monocle's vgam reduced/ null model}
}
\value{
test_res a dataframe containing status of modeling and adjusted p-value
}
\description{
A modified monocle's function for 'compareModels' which identifies and removes genes 
whose reduced_models is better than full_models in term of likelihood
}
\author{
MaiChan Lau
}

