% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSORT_GeneSelection.R
\name{clusterGenes1}
\alias{clusterGenes1}
\title{A modified monocle's function}
\usage{
clusterGenes1(expr_matrix, krange, method = function(x) {     as.dist((1 -
  cor(t(x)))/2) }, ...)
}
\arguments{
\item{expr_matrix}{Expression matrix.}

\item{krange}{krange.}

\item{method}{method function.}

\item{...}{Other parameters.}
}
\value{
test_res a dataframe containing status of modeling and adjusted p-value
}
\description{
A modified monocle's function for 'compareModels' which identifies and removes genes
whose reduced_models is better than full_models in term of likelihood
}
\author{
MaiChan Lau
}

