% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSORT_GeneSelection.R
\name{compareModels1}
\alias{compareModels1}
\title{A modified monocle's function}
\usage{
compareModels1(full_models, reduced_models)
}
\arguments{
\item{full_models}{a Monocle's vgam full model}

\item{reduced_models}{a Monocle's vgam reduced/ null model}
}
\value{
test_res a dataframe containing status of modeling and adjusted p-value
}
\description{
A modified monocle's function for 'compareModels' which identifies and removes genes
whose reduced_models is better than full_models in term of likelihood
}
\author{
MaiChan Lau
}

