% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSORT_GeneSelection.R
\name{diff_test_helper1}
\alias{diff_test_helper1}
\title{A modified monocle's helper function}
\usage{
diff_test_helper1(x, fullModelFormulaStr, reducedModelFormulaStr,
  expressionFamily, lowerDetectionLimit = 0.1, type_ordering = "linear")
}
\arguments{
\item{x}{an expression data}

\item{fullModelFormulaStr}{a Monocle's model structure}

\item{reducedModelFormulaStr}{a Monocle's model structure}

\item{expressionFamily}{a Monocle's family character}

\item{lowerDetectionLimit}{a threshold value}

\item{type_ordering}{a character indicating the type of underlying cell progression,
i.e. linear or circular}
}
\value{
test_res a dataframe containing status of modeling and adjusted p-value
}
\description{
A modified monocle's function for 'diff_test_helper1' which includes more attempts
on finding models and also compute max. magnitude change in expression values predicted
by GLM model
}
\author{
MaiChan Lau
}

