% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSORT_GeneSelection.R
\name{differentialGeneTest1}
\alias{differentialGeneTest1}
\title{differential gene test}
\usage{
differentialGeneTest1(cds,
  fullModelFormulaStr = "expression~sm.ns(Pseudotime, df=3)",
  reducedModelFormulaStr = "expression~1", cores = 1)
}
\arguments{
\item{cds}{Input object.}

\item{fullModelFormulaStr}{Full model formula.}

\item{reducedModelFormulaStr}{Reduced model formula.}

\item{cores}{Number of cores will be used.}
}
\value{
test results
}
\description{
modified from FludigmSC pacakge
}

