% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.function.r
\name{distance.function}
\alias{distance.function}
\title{A distance function
This function allows you to compute cell-to-cell distances}
\usage{
distance.function(expr, flag = 1, method = "eJaccard")
}
\arguments{
\item{expr}{an expresssion matrix containing n-rows of cells and m-cols of genes}

\item{flag}{an integer indicating the type of distance metric: (1) euclidean distance,
(2) correlations, (3) jaccard, (4) input as distance}

\item{method}{a function, a registry entry, or a mnemonic string referencing the 
proximity measure, parameter for \code{proxy::dist} method}
}
\value{
a n-by-n pairwise distance matrix for n cells
}
\description{
A distance function
This function allows you to compute cell-to-cell distances
}
\author{
MaiChan Lau
}

