% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbow_detection_longestDist.r
\name{elbow_detection}
\alias{elbow_detection}
\title{A elbow detection function}
\usage{
elbow_detection(scores = NULL, boundary_cut = 0.1,
  baseNm = "auto.elbow.detection", ylab.text = "Scores",
  xlab.text = "Index")
}
\arguments{
\item{scores}{a data.frame constaining gene names in 'GeneID' column and scores (loading values or
driving forces) in the 'Score' column.}

\item{boundary_cut}{a fraction (at both sides) of the boundary to be omitted for 
differentiating elbow from knee   
the assumed test statistic / distribution of the data. Currently only "Normal" and "CSS" supported.}

\item{baseNm}{a character for naming output files}
}
\value{
a list of selected genes
}
\description{
A elbow detection function detects the elbow/ knee for gene selection, either
using PCA loading values or driving forces. If both elbow & knee are detected
elbow will be returned.
}
\author{
Mai Chan
}

