% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXP_to_CellDataSet.r
\name{EXP_to_CellDataSet}
\alias{EXP_to_CellDataSet}
\title{A data type conversion function}
\usage{
EXP_to_CellDataSet(EXP = NULL, lod = 1)
}
\arguments{
\item{EXP}{a fluidigmSC's EXP object}
}
\value{
a Monocle's CellDataSet object
}
\description{
A data type conversion function transforms fluidigmSC's EXP object to Monocle's 
CellDataSet object.
}
\author{
MaiChan Lau
}

