% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.r
\name{feature_selection}
\alias{feature_selection}
\title{A feature/ gene selection function}
\usage{
feature_selection(cds = NULL, scattering.cutoff = 0.75,
  driving.force.cutoff = NULL, qval_cutoff = NULL, min_expr = 0.1,
  data_type = NULL)
}
\arguments{
\item{cds}{a Monocle's CellDataSet object}

\item{scattering.cutoff}{probability used for removing largely dispersed
genes}

\item{driving.force.cutoff}{a value used for removing genes which do not
change much along cell progress along cell progress path}

\item{qval_cutoff}{a user-defined adjusted p-value below which
genes are retained}

\item{min_expr}{the minimum expression value}

\item{data_type}{a character indicating the type of underlying cell progression, 
i.e. linear or cyclical}
}
\value{
integer
}
\description{
A feature/ gene selection function (1) removes sparsely expressed genes, (2) identifies
differentially expressed genes based on preliminary cell ordering, (3) removes highly
dispersed genes from the identified DEGs, (4) further picks genes which are expected
to have large expression difference on the 2 extreme ends of preliminary cell ordering
}
\author{
MaiChan Lau
}

