% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSort_preProcess.R
\name{fluidigmSC_analyzeGeneDetection}
\alias{fluidigmSC_analyzeGeneDetection}
\title{A gene detection function}
\usage{
fluidigmSC_analyzeGeneDetection(expression_data, threshold = 1)
}
\arguments{
\item{expression_data}{A data frame containing raw expression values, with rownames of genes
and colnames of cells.}

\item{threshold}{A limit of detection in the unit of TPM/CPM/RPMK.}
}
\value{
A data frame containing a column of number of genes detected, and a column of the
corresponding percentage of gene detection, rownames of cells.
}
\description{
A gene detection function computes the fraction of genes detected in each cell,
reproduced from FluidigmSC package.
}

