% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSort_preProcess.R
\name{fluidigmSC_isElementIgnoreCase}
\alias{fluidigmSC_isElementIgnoreCase}
\title{A gene finding function}
\usage{
fluidigmSC_isElementIgnoreCase(x, y, ignore_case = TRUE)
}
\arguments{
\item{x}{A vector of characters representing gene names (target genes).}

\item{y}{A vector of characters representing gene names (source genes).}

\item{ignore_case}{Boolean, if TRUE ignores letter case.}
}
\value{
A vector of characters representing gene names.
}
\description{
A gene finding function looking for genes in the target set x from the source set y,
reproduced from FluidigmSC package.
}

