% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSort_preProcess.R
\name{fluidigmSC_readLinearExp}
\alias{fluidigmSC_readLinearExp}
\title{An expression reading function}
\usage{
fluidigmSC_readLinearExp(exp_file = TRUE, lod = 1)
}
\arguments{
\item{exp_file}{Input file name in txt format, with rownames of cells and colnames of genes.}

\item{lod}{A value of limit of detection in the unit of TPM/CPM/RPKM. It will be used as the starting value
for outlier cell detection and the basis for removing scarce genes.}
}
\value{
A list containing \code{expression_data_raw}(data frame), \code{log2ex_data}(data frame),
and \code{log2ex_avg_data}(data frame).
}
\description{
An expression reading function which imports expression data from .txt file, and then
computes log2 transformed data, reproduced from FluidigmSC package.
}

