% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSORT_preProcess.R
\name{fluidigmSC_removeGenesByLinearExpForAllType}
\alias{fluidigmSC_removeGenesByLinearExpForAllType}
\title{A gene trimming function}
\usage{
fluidigmSC_removeGenesByLinearExpForAllType(log2ex_data, log2ex_avg_data,
  threshold)
}
\arguments{
\item{log2ex_data}{A data frame containing log2 tranformed expression values, with rownames of genes
and colnames of cells.}

\item{log2ex_avg_data}{A data frame containing log2 tranformed average expression values for individual
gene.}

\item{threshold}{A limit of detection in the unit of TPM/CPM/RPMK.}
}
\value{
A vector of character containing gene names of those passed the filtering.
}
\description{
A gene trimming function removes genes whose average expression value is below the log2(threshold),
and also present in at least 10% of total cells;  reproduced from FluidigmSC package.
}

