% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSort_preProcess.R
\name{fluidigmSC_removeGenesByLinearExpForAllType_log2}
\alias{fluidigmSC_removeGenesByLinearExpForAllType_log2}
\title{A gene trimming function}
\usage{
fluidigmSC_removeGenesByLinearExpForAllType_log2(log2ex_data, threshold)
}
\arguments{
\item{log2ex_data}{A data frame containing log2 tranformed expression values, with rownames of genes
and colnames of cells.}

\item{threshold}{A limit of detection in the unit of TPM/CPM/RPMK.}
}
\value{
A vector of character containing gene names of those passed the filtering.
}
\description{
A gene trimming function removes genes whose average expression value is below the log2(threshold);
reproduced from FluidigmSC package.
}

