% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_cluster_GO.r
\name{gene_cluster_GO}
\alias{gene_cluster_GO}
\title{A gene ontology function}
\usage{
gene_cluster_GO(resDir = NULL, GO_controls = NULL,
  universal_geneSet = NULL, geneClusterID_fname = NULL, baseNm = "GO")
}
\arguments{
\item{resDir}{the directory to which output results will be written}

\item{GO_controls}{parameters for gene enrichment study}

\item{universal_geneSet}{a list of universal or backgroupd genes}

\item{gene.clusterID}{a dataframe consists of a column 'GeneID' for a list of
genes & a column of 'ClusterID' for the corresponding cluster assignment of
each gene}
}
\description{
A gene ontology function performs gene set enrichment analysis using topGO package
}
\author{
MaiChan Lau
}

