% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_cluster_heatmap.r
\name{gene_cluster_heatmap}
\alias{gene_cluster_heatmap}
\title{A heatmap plotting function}
\usage{
gene_cluster_heatmap(resDir = NULL, sorted_exp = NULL,
  gene.clusterID = NULL, flipping = FALSE)
}
\arguments{
\item{resDir}{the directory to which output plot will be stored}

\item{sorted_exp}{a EXP object with cells sorted}

\item{gene.clusterID}{a dataframe consists of a column 'GeneID' for a list of
genes & a column of 'ClusterID' for the corresponding cluster assignment of
each gene}

\item{flipping}{a logical value stating if the final cell ordering is flipped
for heatmap plotting}
}
\description{
A heatmap plotting function
}
\author{
MaiChan Lau
}

