% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_cluster_profile.r
\name{gene_cluster_profile}
\alias{gene_cluster_profile}
\title{A function generating gene subsets}
\usage{
gene_cluster_profile(exp = NULL, cell.ordering = NULL, geneCluster = NULL,
  resDir = NULL)
}
\arguments{
\item{exp}{an EXP object}

\item{cell.ordering}{a data.frame of cell ordering; in the format of 
sample_list of EXP object}

\item{geneCluster}{a data.frame containing clusterID of a set of interesting genes}

\item{resDir}{an output directory}
}
\description{
A function generating gene subsets
}
\author{
MaiChan Lau
}

