% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyOutliers_noSave.r
\name{identifyOutliers}
\alias{identifyOutliers}
\title{A modified function of fluidigmSC's identifyOutliers}
\usage{
identifyOutliers(exp_file = NULL, lod = 1)
}
\arguments{
\item{exp_file}{a character of input expression filename}

\item{lod}{limit of detection used for gene trimming}
}
\value{
The updated EXP object
}
\description{
A modified version of fluidigmSC's identifyOutliers function which removes the saving
feature.
}
\author{
fluidigmSC
}

