% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_sorting.r
\name{neighborhood_sorting}
\alias{neighborhood_sorting}
\title{A sorting function using the neighborhood algorithm}
\usage{
neighborhood_sorting(d, weights_mat = NULL, max_iter = 100)
}
\arguments{
\item{d}{an expresssion matrix containing n-by-n pairwise distances computed for n cells}

\item{max_iter}{maximum number of iteration}
}
\value{
a list containinig a vector \code{ordering} of the permutated cell ordering, and a cost value \code{cost}
}
\description{
This function allows you to order cells according to the neighborhood
algorithm proposed in SPIN paper Tsafrir, D. et al. (2005). Bioinformatics, 21(10), page 2301 to 2308.
}
\author{
MaiChan Lau
}

