% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_sorting_wrapper.r
\name{neighborhood_sorting_wrapper}
\alias{neighborhood_sorting_wrapper}
\title{A neighborhood sorting wrapper function}
\usage{
neighborhood_sorting_wrapper(expr, sigma_width = 1, no_randomization = 10)
}
\arguments{
\item{expr}{an expresssion matrix containing n-rows of cells and m-cols of genes}

\item{repeats}{number of randomly chosen starting ordering to be tested}
}
\value{
a list containinig a matrix \code{permutated.expr} of the permutated expression matrix
and the best cost value \code{best.cost}
}
\description{
A sorting wrapper function seeks for an (locally) optimal cell permutation by 
repeating the neighborhood sorting operation (i.e. neighborhood_sorting) on randomly chosen starting ordering
}
\author{
MaiChan Lau
}

