% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSORT_GeneSelection.R
\name{pca_gene_selection}
\alias{pca_gene_selection}
\title{Gene selection using PCA technique}
\usage{
pca_gene_selection(data)
}
\arguments{
\item{data}{A matrix of data.frame with row.name of cells, and col.name of genes}
}
\value{
a vector of the names of selected genes.
}
\description{
Gene selection using PCA technique
}
\examples{
dir <- system.file('extdata', package='uSORT')
file <- list.files(dir, pattern='.txt$', full=TRUE)
exprs <- uSORT_preProcess(exprs_file = file)
exp_trimmed <- t(exprs$exprs_log_trimed)
PCA_selected_genes <- pca_gene_selection(exp_trimmed)
}

