% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.r
\name{pre_processing}
\alias{pre_processing}
\title{A data pre-processing function}
\usage{
pre_processing(exp = NULL)
}
\arguments{
\item{exp}{a fluidigmSC's EXP object output from identifyOutliers function}
}
\value{
a fluidigmSC's EXP object which is ready for downstream analysis
}
\description{
A data pre-processing function removes genes which are lowly expressed and performs PCA analysis
based on FluidigmSC's autoAnalysis function.
}
\author{
MaiChan Lau
}

