% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSORT_GeneSelection.R
\name{scattering_quantification_per_gene}
\alias{scattering_quantification_per_gene}
\title{An expression scattering measurement function}
\usage{
scattering_quantification_per_gene(CDS = NULL)
}
\arguments{
\item{CDS}{a Monocle's CellDataSet object}
}
\value{
integer
}
\description{
An expression scattering measurement function computes the level of scattering
for individual genes along the cell ordering
}
\author{
MaiChan Lau
}

