% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSORT_GUI.R
\name{sorting_method_parameter_GUI}
\alias{sorting_method_parameter_GUI}
\title{GUI for sorting method paramters}
\usage{
sorting_method_parameter_GUI(method = c("autoSPIN", "sWanderlust", "monocle",
  "Wanderlust", "SPIN", "none"))
}
\arguments{
\item{method}{method name.}
}
\value{
a list of parameters.
}
\description{
The parameters appeared on GUI are based on input method, this function is called by
\code{\link{uSORT_parameters_GUI}}. For internal use only.
}
\author{
Hao Chen
}

