% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_wrapper.r
\name{sorting_wrapper}
\alias{sorting_wrapper}
\title{A sorting wrapper function}
\usage{
sorting_wrapper(expr, no_randomization = 10, local_sort = FALSE,
  window_perc_range = c(0.1, 0.9), alpha = 0.2,
  window_size_incre_perct = 0.05, sorting_method = NULL, baseNm = "uSPIN",
  sigma_width = 1, data_type = NULL)
}
\arguments{
\item{expr}{an expresssion matrix containing n-rows of cells and m-cols of genes}

\item{local_sort}{a logical stating if local sorting feature is implemented: (1) turn on, 
(0) turn off}

\item{window_perc_range}{range of window size given in percentages of total no. of cells 
indicating the smallest and largest window sizes for local sorting refinement}

\item{alpha}{window size which is given in  percentage of total no. of cells,
parameter for \code{c2c_dist_uneveness}}

\item{window_size_incre_perct}{size increment in percentage for local sorting refinement}

\item{sorting_method}{character specifying the type of sorthing algorithm, i.e. STS or neighborhood}

\item{baseNm}{base name used for saving output plot}

\item{repeats}{number of randomly chosen starting ordering to be tested, 
parameter for \code{STS_sorting_wrapper} method}
}
\value{
a matrix of the permutated expression matrix
}
\description{
A sorting wrapper function includes 2 additional features for the original SPIN method, 
(1) optimization with multiple randomly chosen starting ordering in 
\code{STS_sorting_wrapper}, and (2) local sorting refinement through 
\code{c2c_dist_uneveness}
}
\author{
MaiChan Lau
}

