% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STS_sorting.r
\name{STS_sorting}
\alias{STS_sorting}
\title{A sorting function using the Side-to-side (STS) algorithm}
\usage{
STS_sorting(d, max_iter = 10)
}
\arguments{
\item{d}{an expresssion matrix containing n-by-n pairwise distances computed for n cells}

\item{max_iter}{maximum number of iteration}
}
\value{
a list containinig a vector \code{ordering} of the permutated cell ordering, and a cost value \code{cost}
}
\description{
This function allows you to order cells according to the Side-to-side (STS) 
algorithm proposed in SPIN paper Tsafrir, D. et al. (2005). Bioinformatics, 21(10), page 2301 to 2308.
}
\author{
MaiChan Lau
}

