% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STS_sorting_wrapper.r
\name{STS_sorting_wrapper}
\alias{STS_sorting_wrapper}
\title{A STS sorting wrapper function}
\usage{
STS_sorting_wrapper(expr, no_randomization = 10)
}
\arguments{
\item{expr}{an expresssion matrix containing n-rows of cells and m-cols of genes}

\item{repeats}{number of randomly chosen starting ordering to be tested}
}
\value{
a list containinig a matrix \code{permutated.expr} of the permutated expression matrix
and the best cost value \code{best.cost}
}
\description{
A sorting wrapper function seeks for an (locally) optimal cell permutation by 
repeating the STS sorting operation (i.e. STS_sorting) on randomly chosen starting ordering
}
\author{
MaiChan Lau
}

