% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoSPIN.R
\name{summed_local_variance_linear}
\alias{summed_local_variance_linear}
\title{A summed local variance function for strictly linear data type}
\usage{
summed_local_variance_linear(d, alpha = 0.3)
}
\arguments{
\item{d}{A cell-to-cell distance matrix.}

\item{alpha}{A fraction value indicating the size of window for local variance measurement.}
}
\value{
A numeric value of the summed local variance.
}
\description{
A summed local variance function for strictly linear data type
}

