% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topGO_geneCluster.r
\name{topGO_geneCluster}
\alias{topGO_geneCluster}
\title{A gene enrichment function}
\usage{
topGO_geneCluster(geneList = NULL, GO_controls = NULL, resDir = NULL,
  baseNm = "GO")
}
\arguments{
\item{geneList}{a data.frame of gene names and cluster assignment}

\item{GO_controls}{a data.frame containing the choice of the ontology type & mapping}

\item{resDir}{the directory to which output results will be written}
}
\description{
A gene enrichment function which is based on topGO package. It identifies enriched function 
of each gene clusters.
}
\author{
MaiChan Lau
}

