% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swanderlust.R
\name{trajectory_landmarks}
\alias{trajectory_landmarks}
\title{determining initial trajectory and landmarks}
\usage{
trajectory_landmarks(knn, data, s, partial_order = NULL, waypoints = 250,
  waypoints_seed = 123, metric = "euclidean", flock_waypoints = 2,
  band_sample = FALSE)
}
\arguments{
\item{knn}{A sparse matrix of knn.}

\item{data}{data.}

\item{s}{The ID of starting point.}

\item{partial_order}{A vector of IDs specified as recommended waypoints, NULL to ignore.}

\item{waypoints}{Either the number of waypoints, or specify the waypoint IDs.}

\item{waypoints_seed}{Random sampling seed, for reproducible results.}

\item{metric}{Distance calculation metric for nearest neighbour detection.}

\item{flock_waypoints}{Iteration of using nearest points around waypoint to adjust its position.}

\item{band_sample}{if give more chance to nearest neighbours of starting point in randomly waypoints selection.}
}
\value{
a list
}
\description{
determining initial trajectory and landmarks
}

