% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSort.r
\docType{package}
\name{uSort}
\alias{uSort}
\alias{uSort-package}
\title{uSort: A self-refining ordering pipeline for gene selection}
\description{
This package is designed to uncover the intrinsic cell progression path
from single-cell RNA-seq data.
}
\details{
This package incorporates data pre-processing, preliminary PCA gene selection, preliminary
cell ordering, feature selection, refined cell ordering, and post-analysis interpretation
and visualization. The uSort workflow can be implemented through calling the main function
\code{\link{uSort_main}}.
}

