% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSORT_GUI.R
\name{uSORT_GUI}
\alias{uSORT_GUI}
\title{The user friendly GUI for \code{uSORT-package}}
\usage{
uSORT_GUI()
}
\value{
the GUI for \code{uSORT-package}
}
\description{
This GUI provides an easy way for applying the uSORT package.
}
\examples{
interactive()
#if(interactive()) uSORT_GUI()  # remove the hash symbol to run
}
\author{
Hao Chen
}
\references{
\url{http://JinmiaoChenLab.github.io/uSORT/}
}
\seealso{
\code{\link{uSORT-package}}, \code{\link{uSORT}}
}

