% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSort.r
\name{uSort_main}
\alias{uSort_main}
\title{A main function of uSort}
\usage{
uSort_main(resDir = NULL, exprsFile = NULL, scattering.cutoff = 0.75,
  exp.outlier.removed = NULL, driving.force.cutoff = NULL,
  data_type = NULL, qval_cutoff_featureselection = 0.05, sigma_width = 1,
  sorting_method = NULL, no_randomization = NULL, alpha = NULL)
}
\arguments{
\item{resDir}{the directory to which output results will be written}

\item{exprsFile}{a character of input expression filename}

\item{exp.outlier.removed}{a EXP object of preprocessed expression data}

\item{driving.force.cutoff}{a value used for removing genes which do not
change much along cell progress}

\item{data_type}{a character indicating the type of underlying cell progression, 
i.e. linear or cyclical}

\item{qval_cutoff_featureselection}{a user-defined adjusted p-value below which
genes are selected for driver gene selection}
}
\value{
a matrix \code{res} of the permutated norm_exprs. If choose 'writeResults = TRUE', results
will be saved into files under \code{resDir}
}
\description{
A main function of uSort \code{uSort_main} which provides a workflow of sorting scRNA-seq data,
including data preprocess with merging methods of multiple fcs file, logicle transformation,
dimension reduction with PCA, isomap or tsne(default), and a kernal-based local maxima
clustering combined with SVM for subpopulation detection. The intermediate results can be saved
into seperate files and the cluster results can be visualized in heatmaps and scatter plots.
}
\author{
MaiChan Lau
}

