% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSORT_GUI.R
\name{uSORT_parameters_GUI}
\alias{uSORT_parameters_GUI}
\title{The GUI for inputting paramters for uSORT}
\usage{
uSORT_parameters_GUI()
}
\value{
a list of parameters.
}
\description{
This is a function for generating the GUI for uSORT,
it's called by \code{\link{uSORT_GUI}}. For internal use only.
}
\author{
Hao Chen
}

