% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSORT_postProcess.R
\name{uSORT_write_results}
\alias{uSORT_write_results}
\title{Resluts parsing for uSORT}
\usage{
uSORT_write_results(uSORT_results, project_name, result_directory)
}
\arguments{
\item{uSORT_results}{Result object from \code{uSort} function, a list.}

\item{project_name}{A prefix for the saving files.}

\item{result_directory}{The path where to save the results.}
}
\value{
save the results.
}
\description{
Save result object into a RData file. Save cell to cell distance heatmap for
both preliminary and refined results. Creat plot of driver gene profiles on
final ordering using heatmap.
}
\examples{
dir <- system.file('extdata', package='uSORT')
file <- list.files(dir, pattern='.txt$', full=TRUE)
#remove the # symbol of the following codes to test
#uSORT_results <- uSORT(exprs_file = file,
# project_name = 'test',
# preliminary_sorting_method = 'autoSPIN',
# refine_sorting_method = 'sWanderlust',
# save_results = FALSE)
#uSORT_write_results(uSORT_results,
# project_name = 'test',
# result_directory = getwd())
}

