% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSORT_GeneSelection.R
\name{variability_per_gene}
\alias{variability_per_gene}
\title{A utility function for scattering_quantification_per_gene}
\usage{
variability_per_gene(logExp = NULL, min_expr = 0.1,
  window_size_perct = 0.1, nonZeroExpr_perct = 0.1)
}
\arguments{
\item{logExp}{a log-scale expression vector of a gene}

\item{min_expr}{a minimum expression value}

\item{window_size_perct}{a window size (in % of total no. of cells) used for computing
dispersion level}

\item{nonZeroExpr_perct}{a minimum amount of cells (in % of window size) with non-zero
expression, otherwise the associated window will be assigned to 0 disperson value}
}
\value{
integer
}
\description{
A utility function for scattering_quantification_per_gene which computes the degree
of scattering for single gene, whereby the value is computed by summing over the
local values of smaller local windows
}
\author{
MaiChan Lau
}

