% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swanderlust.R
\name{wanderlust_wrapper}
\alias{wanderlust_wrapper}
\title{a wrapper of wanderlust for sWanderlust}
\usage{
wanderlust_wrapper(data, s, diffusionmap_components = 4, l = 15, k = 15,
  num_graphs = 1, num_waypoints = 150, waypoints_seed = 123,
  flock_waypoints = 2)
}
\arguments{
\item{data}{Input data matrix.}

\item{s}{The ID of starting point.}

\item{diffusionmap_components}{Number of components from diffusion map used for wanderlust analysis, default is 4.}

\item{l}{Number of nearest neighbors, default is 15.}

\item{k}{Number of nearest neighbors for repeating graphs, default is 15, should be less than or equal to l.}

\item{num_graphs}{Number of repreated graphs.}

\item{num_waypoints}{Number of waypoint used for wanderlust, default is 150.}

\item{waypoints_seed}{The seed for reproducing the results.}

\item{flock_waypoints}{The number of times for flocking the waypoints, default is 2.}
}
\value{
sorted order.
}
\description{
a wrapper of wanderlust for sWanderlust
}
\author{
Hao Chen
}

