% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoSPIN.R
\name{distance.function}
\alias{distance.function}
\title{A distance function
A distance function computes cell-to-cell distance matrix.}
\usage{
distance.function(expr, method = c("Euclidean", "Correlation", "eJaccard",
  "none"))
}
\arguments{
\item{expr}{An expresssion matrix containing n-rows of cells and m-cols of genes.}

\item{method}{A character string indicating the distance function.}
}
\value{
A matrix containing n-by-n cell distance.
}
\description{
A distance function
A distance function computes cell-to-cell distance matrix.
}

