% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSORT_GeneSelection.R
\name{elbow_detection}
\alias{elbow_detection}
\title{A elbow detection function}
\usage{
elbow_detection(scores, if_plot = FALSE)
}
\arguments{
\item{scores}{A vector of numeric scores.}

\item{if_plot}{Boolean determine if plot the results.}
}
\value{
a vector of selected elements IDs
}
\description{
A elbow detection function detects the elbow/knee of a given vector of values.
Values will be sorted descendingly before detection, and the ID of those values
above the elbow will be returned.
}
\examples{
scores <- c(10, 9 ,8, 6, 3, 2, 1, 0.1)
elbow_detection(scores, if_plot = TRUE)
}

