% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle.R
\name{EXP_to_CellDataSet}
\alias{EXP_to_CellDataSet}
\title{A function for constructing a Monocle's CellDataSet object from an expression matrix}
\usage{
EXP_to_CellDataSet(log2_exp = NULL, expression_data_raw = NULL, lod = 1)
}
\arguments{
\item{log2_exp}{An log2 transformed expresssion matrix containing n-rows of cells and m-cols of genes.}

\item{expression_data_raw}{A data frame containing raw expression values, with rownames of cells
and colnames of genes.}

\item{lod}{A value of limit of detection in the unit of TPM/CPM/RPKM.}
}
\value{
A CellDataSet object.
}
\description{
A function for constructing a Monocle's CellDataSet object from an expression matrix
}

