% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIN.R
\name{neighborhood_sorting}
\alias{neighborhood_sorting}
\title{A sorting function using the Neighborhood algorithm}
\usage{
neighborhood_sorting(d, weights_mat = NULL, max_iter = 100)
}
\arguments{
\item{d}{A matrix containing n-by-n cell distance.}

\item{weights_mat}{A weight matrix of size n-by-n.}

\item{max_iter}{An integer number indicating the maximum number of iteration if sorting
does not converge.}
}
\value{
A list containing \code{ordering}(a vector of re-ordered sequence) and
\code{cost}(a numeric value).
}
\description{
A sorting function using the Neighborhood algorithm
}

