% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIN.R
\name{neighborhood_sorting_wrapper}
\alias{neighborhood_sorting_wrapper}
\title{A wrapper function for Neighborhood sorting.}
\usage{
neighborhood_sorting_wrapper(expr, sigma_width = 1, no_randomization = 10)
}
\arguments{
\item{expr}{An expresssion matrix containing n-rows of cells and m-cols of genes.}

\item{sigma_width}{An integer number determining the degree of spread of the gaussian distribution which is used for computing weight matrix for Neighborhood sorting method.}

\item{no_randomization}{An integer number indicating the number of repeated sorting, each of which uses a randomaly selected initial cell ordering.}
}
\value{
A list containing \code{permutated.expr}(data frame) and \code{best.cost}(a numeric value).
}
\description{
A wrapper function for Neighborhood sorting as proposed in [Tsafrir et al. 2005].
}

