% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIN.R
\name{STS_sorting}
\alias{STS_sorting}
\title{A sorting function using the Side-to-Side (STS) algorithm}
\usage{
STS_sorting(d, max_iter = 10)
}
\arguments{
\item{d}{A matrix containing n-by-n cell distance.}

\item{max_iter}{An integer number indicating the maximum number of iteration if sorting
does not converge.}
}
\value{
A list containing \code{ordering}(a vector of re-ordered sequence) and
\code{cost}(a numeric value).
}
\description{
A sorting function using the Side-to-Side (STS) algorithm
}

