% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIN.R
\name{STS_sorting_wrapper}
\alias{STS_sorting_wrapper}
\title{A wrapper function for Side-to-Side (STS) sorting.}
\usage{
STS_sorting_wrapper(expr, no_randomization = 10)
}
\arguments{
\item{expr}{An expresssion matrix containing n-rows of cells and m-cols of genes.}

\item{no_randomization}{An integer number indicating the number of repeated sorting, each
of which uses a randomaly selected initial cell ordering.}
}
\value{
A list containing \code{permutated.expr}(data frame) and \code{best.cost}(a numeric value).
}
\description{
A wrapper function for Side-to-Side (STS) sorting as proposed in [Tsafrir et al. 2005].
}

